requirements.txt
faiss-cpu
langchain
langchain-ollama
pytesseract
pdfplumber
pillow
numpy
collections
streamlit
csv
